refactor modelrunner
